Bad Hack to not specify endpoint in web.config
This is a very bad hack.  I needed a way to specify an endpoint by not
putting it in the web.config.  I have a web application where the
customer can enable/disable the test (sandbox) mode while the
application is running.  I did not want this to be tied to a web.config
option.  Since I was specifying the credentials each call, I decided to
put the endpoint in SignatureCredentials even though I understand this
is not the correct place for it.  I am still developing this
application, but for the time being this seems to have worked for me.  I
have not broken any existing functionality, just slightly extended it.